moved FavoriteStyles.ts to Favorites folder
