+++ src/pages/Favorites/FavoriteStyles.ts
